Rename: #12 컨트롤러, 서비스, DTO 위치 및 이름 정리
