+++ backend/src/main/java/com/ohammer/apartner/domain/facility/controller/FacilityUserController.java
@@ -1,7 +1,7 @@
 package com.ohammer.apartner.domain.facility.controller;
 
-import com.ohammer.apartner.domain.facility.dto.FacilityResponseDto;
-import com.ohammer.apartner.domain.facility.service.FacilityService;
+import com.ohammer.apartner.domain.facility.dto.response.FacilityResponseDto;
+import com.ohammer.apartner.domain.facility.service.FacilityUserService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -14,11 +14,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/facilities")
 public class FacilityController {
 
-    private final FacilityService facilityService;
+    private final FacilityUserService facilityUserService;
 
     @GetMapping
     public ResponseEntity<List<FacilityResponseDto>> getAllFacilities() {
-        List<FacilityResponseDto> facilities = facilityService.getAllFacilities();
+        List<FacilityResponseDto> facilities = facilityUserService.getAllFacilities();
         return ResponseEntity.ok(facilities);
     }
 }
